test 00

+++ include/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: kadam <kadam@student.1337.ma>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/04 12:12:37 by kadam             #+#    #+#             */
-/*   Updated: 2024/09/06 10:04:18 by kadam            ###   ########.fr       */
+/*   Updated: 2024/09/06 11:34:13 by kadam            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ int ft_lenarray(char **array);
 void free_2d_array(char **array);
 void ft_putstrn_fd(char *s, int fd);
 char **ft_split(char *s, char *c);
-char *ft_substr(char const *s, unsigned int start, size_t len);
+char *ft_substr(char const *s, size_t start, size_t len);
 char *ft_strtrim(char const *s1, char const *set);
 int ft_strcmp(const char *s1, const char *s2);
 int ft_atoi(const char *str);
